Create particle_commander_v0.1.py
The Commander script written in Python 3 is designed to create a set of particle spawning files which is designed for particle tracking in the GETM/GOTM Cefas 3D Hydrodynamic, Meteorological and Biogeochemical oceanographic model. 

It can be adapted for use with other models that want to create similar particle spawning files. 

I would advise any model that would intend to add particle spawning to apply a standard to the spawn_input.file:
14 {number of particles}
1, -10.300000000000001, 48.84999999999998, 0.0, '2001/01/01 00:59:60', 1000
2, -10.300000000000001, 48.84999999999998, 10.0, '2001/01/01 00:59:60', 1000
{particle number}, {latitude}, {longitude}, {depth}, {date & start time}, 

The Commanding script currently calls two other scripts - particle_spawn_v0.1.py (python) and lag_getm.inp (fortran). 
It uses the above two as templates and creates two new set of files.

The particle_spawn_script is used to create a particle_spawn_file with X amount of particles at X lon and Y lat 
The lag_getm.inp calls the particle_spawn_script using the data within as start parameters for the particles